+++ app/components/atoms/Button/LinkButton.js
@@ -3,19 +3,19 @@ import styled from 'styled-components';
 
 export const LinkButton = styled(NavLink)`
   display: inline-flex;
-  padding: 32px 24px;
+  padding: calc(var(--spacing-factor) * 3);
   text-decoration: none;
   -webkit-font-smoothing: antialiased;
   -webkit-touch-callout: none;
   user-select: none;
   cursor: pointer;
   outline: 0;
-  font-family: 'Helvetica Neue', Helvetica, Arial, sans-serif;
-  font-weight: bold;
-  font-size: 16px;
+  font-size: 1.1em;
   color: var(--light);
   position: relative;
   justify-content: center;
+  text-transform: uppercase;
+  font-family: var(--font-family-title);
 
   &:after {
     content: '';